requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
certifi==2025.1.31
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
rich==14.0.0
sniffio==1.3.1
textual==3.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uc-micro-py==1.0.3
